admin: refacto information section into edit component
