+++ admin/app/components/organizations/information-section.hbs
@@ -13,145 +13,12 @@
     </div>
 
     {{#if this.isEditMode}}
-      <div class="organization__edit-form">
-        <form class="form" {{on "submit" this.updateOrganization}}>
-
-          <span class="form__instructions">
-            Les champs marqués de
-            <abbr title="obligatoire" class="mandatory-mark" aria-hidden="true">*</abbr>
-            sont obligatoires.
-          </span>
-
-          <div class="form-field">
-            <label for="name" class="form-field__label">
-              <abbr title="obligatoire" class="mandatory-mark" aria-hidden="true">*</abbr>
-              Nom
-            </label>
-            {{#if (v-get this.form "name" "isInvalid")}}
-              <div class="form-field__error">
-                {{v-get this.form "name" "message"}}
-              </div>
-            {{/if}}
-            <Input
-              id="name"
-              @type="text"
-              class={{if (v-get this.form "name" "isInvalid") "form-control is-invalid" "form-control"}}
-              @value={{this.form.name}}
-              required={{true}}
-            />
-          </div>
-          <div class="form-field">
-            <label for="externalId" class="form-field__label">Identifiant externe</label>
-            {{#if (v-get this.form "externalId" "isInvalid")}}
-              <div class="form-field__error">
-                {{v-get this.form "externalId" "message"}}
-              </div>
-            {{/if}}
-            <Input
-              id="externalId"
-              @type="text"
-              class={{if (v-get this.form "externalId" "isInvalid") "form-control is-invalid" "form-control"}}
-              @value={{this.form.externalId}}
-            />
-          </div>
-          <div class="form-field">
-            <label for="provinceCode" class="form-field__label">Département (en 3 chiffres)</label>
-            {{#if (v-get this.form "provinceCode" "isInvalid")}}
-              <div class="form-field__error">
-                {{v-get this.form "provinceCode" "message"}}
-              </div>
-            {{/if}}
-            <Input
-              id="provinceCode"
-              @type="text"
-              class={{if (v-get this.form "provinceCode" "isInvalid") "form-control is-invalid" "form-control"}}
-              @value={{this.form.provinceCode}}
-            />
-          </div>
-          <div class="form-field">
-            <label for="email" class="form-field__label">Adresse e-mail d'activation SCO</label>
-            {{#if (v-get this.form "email" "isInvalid")}}
-              <div class="form-field__error">
-                {{v-get this.form "email" "message"}}
-              </div>
-            {{/if}}
-            <Input
-              id="email"
-              @type="text"
-              class={{if (v-get this.form "email" "isInvalid") "form-control is-invalid" "form-control"}}
-              @value={{this.form.email}}
-            />
-          </div>
-          <div class="form-field">
-            <label for="credits" class="form-field__label">Crédits</label>
-            {{#if (v-get this.form "credit" "isInvalid")}}
-              <div class="form-field__error">
-                {{v-get this.form "credit" "message"}}
-              </div>
-            {{/if}}
-            <Input
-              id="credits"
-              @type="number"
-              class={{if (v-get this.form "credit" "isInvalid") "form-control is-invalid" "form-control"}}
-              @value={{this.form.credit}}
-            />
-          </div>
-          <div class="form-field">
-            <label for="documentationUrl" class="form-field__label">Lien vers la documentation</label>
-            {{#if (v-get this.form "documentationUrl" "isInvalid")}}
-              <div class="form-field__error">
-                {{v-get this.form "documentationUrl" "message"}}
-              </div>
-            {{/if}}
-            <Input
-              id="documentationUrl"
-              @type="text"
-              class={{if (v-get this.form "documentationUrl" "isInvalid") "form-control is-invalid" "form-control"}}
-              @value={{this.form.documentationUrl}}
-            />
-          </div>
-          {{#if (or @organization.isOrganizationSCO @organization.isOrganizationSUP)}}
-            <div class="form-field organization-edit-form__checkbox">
-              <label for="isManagingStudents" class="form-field__label">Gestion d’élèves/étudiants</label>
-              {{#if (v-get this.form "isManagingStudents" "isInvalid")}}
-                <div class="form-field__error">
-                  {{v-get this.form "isManagingStudents" "message"}}
-                </div>
-              {{/if}}
-              <Input
-                id="isManagingStudents"
-                @type="checkbox"
-                class={{if (v-get this.form "isManagingStudents" "isInvalid") "form-control is-invalid" "form-control"}}
-                @checked={{this.form.isManagingStudents}}
-              />
-            </div>
-          {{/if}}
-          <div class="form-field organization-edit-form__checkbox">
-            <label for="showSkills" class="form-field__label">Affichage des acquis dans l'export de résultats</label>
-            {{#if (v-get this.form "showSkills" "isInvalid")}}
-              <div class="form-field__error">
-                {{v-get this.form "showSkills" "message"}}
-              </div>
-            {{/if}}
-            <Input
-              id="showSkills"
-              @type="checkbox"
-              class={{if (v-get this.form "showSkills" "isInvalid") "form-control is-invalid" "form-control"}}
-              @checked={{this.form.showSkills}}
-            />
-          </div>
-          <div class="form-actions">
-            <PixButton
-              @size="small"
-              @backgroundColor="transparent-light"
-              @isBorderVisible={{true}}
-              @triggerAction={{this.cancel}}
-            >Annuler</PixButton>
-            <PixButton @type="submit" @size="small" @backgroundColor="green">Enregistrer</PixButton>
-          </div>
-        </form>
-      </div>
-
+      <Organizations::Information-section-edition-mode
+        @organization={{@organization}}
+        @toggleEditMode={{this.toggleEditMode}}
+        @cancel={{this.cancel}}
+        @onSubmit={{@onSubmit}}
+      />
     {{else}}
       <Organizations::Information-section-read-mode
         @organization={{@organization}}
